+++ src/Pim/Bundle/EnrichBundle/Manager/SequentialEditManager.php
@@ -56,6 +56,8 @@ class SequentialEditManager implements SaverInterface, RemoverInterface
 
     /**
      * {@inheritdoc}
+     *
+     * TODO: extract to saver
      */
     public function save($object, array $options = [])
     {
@@ -87,6 +89,8 @@ class SequentialEditManager implements SaverInterface, RemoverInterface
 
     /**
      * {@inheritdoc}
+     *
+     * TODO: extract to remover
      */
     public function remove($object, array $options = [])
     {