+++ CHANGELOG.md
@@ -128,6 +128,7 @@
 - Constructor of `Pim\Bundle\EnrichBundle\Controller\ChannelController` now takes a BulkSaver as last argument (to save locales)
 - Constructor of `Pim\Bundle\CatalogBundle\Manager\AttributeManager` has been changed
 - Constructor of `Pim\Bundle\CatalogBundle\Manager\AttributeGroupManager` has been changed
+- Constructor of `Pim\Bundle\CommentBundle\Manager\CommentManager` has been changed
 
 # 1.3.x